anamnesia/requirements.txt
flask==3.1.0
flask-cors==5.0.0
python-dotenv==1.0.1
supabase==2.10.0
groq==0.13.0
faster-whisper==1.1.1
torch==2.6.0
torchaudio==2.6.0
gunicorn==23.0.0
PyJWT[crypto]
